feat: create trip detail
