+++ package.json
@@ -43,12 +43,12 @@
     "react-native-linear-gradient": "^2.8.3",
     "react-native-maps": "1.18.0",
     "react-native-pager-view": "^6.7.1",
-    "react-native-modal": "^14.0.0-rc.1",
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-snap-carousel": "^3.9.1",
     "react-native-snap-carousel-v4": "^1.0.1",
+    "react-native-svg": "^15.12.0",
     "react-native-toast-message": "^2.3.0",
     "react-native-web": "~0.19.13"
   },
